src/app/components/sidebar-menu/sidebar-menu.component.ts
import { Component, ViewChild, AfterViewInit } from '@angular/core';
import { MatSidenav } from '@angular/material/sidenav';


@Component({
  selector: 'app-sidebar-menu',
  templateUrl: './sidebar-menu.component.html',
  styleUrls: ['./sidebar-menu.component.css'],
})
export class SidebarMenuComponent implements AfterViewInit {
  @ViewChild('sidenav', { static: false }) sidenav!: MatSidenav;

  constructor() {}

  ngAfterViewInit() {
  }

  openSidenav() {
    this.sidenav.open();
  }

  closeSidenav() {
    this.sidenav.close();
  }
}
